move report maker to pipeline.publish
